Added a fly-through asteroid
I have the FBXs to the other asteroids in the scene, just need to get the textures on and place them. It's late so That'll have to wait until tomorrow. For now I cleaned up the scene and made sure the main asteroid was in so I can build around that.